Add counter units, decimals, templates, and Health-style history.
Ship the remaining product gaps: optional units in the header, hundredths-based decimals, create templates, list reorder/delete, CSV export, and a paginated history chart with hourly day buckets and whole-number axis labels.

+++ Counter/Views/Components/CustomAmountSheet.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
 struct CustomAmountSheet: View {
-  let onAdd: (Int) -> Void
+  let onAdd: (Double) -> Void
 
   var body: some View {
     AmountEntrySheet(
