+++ Shared/QuickAddConfiguration.swift
@@ -4,14 +4,14 @@ nonisolated enum QuickAddConfiguration {
   static let buttonCount = 10
   static let presetCount = buttonCount - 1
 
-  static let defaultCaloriePresets: [Int] = [5, 10, 25, 50, 100, 200, 500, 1000, 50]
-  static let defaultCounterPresets: [Int] = [1, 2, 5, 10, 20, 50, 100, 25, 75]
+  static let defaultCaloriePresets: [Double] = [5, 10, 25, 50, 100, 200, 500, 1000, 50]
+  static let defaultCounterPresets: [Double] = [1, 2, 5, 10, 20, 50, 100, 25, 75]
 
-  static func normalizedPresets(_ values: [Int]) -> [Int] {
+  static func normalizedPresets(_ values: [Double]) -> [Double] {
     Array(values.sorted().prefix(presetCount))
   }
 
-  static func filledPresets(from stored: [Int], defaults: [Int]) -> [Int] {
+  static func filledPresets(from stored: [Double], defaults: [Double]) -> [Double] {
     var result = normalizedPresets(stored)
     guard result.count < presetCount else { return result }
 
@@ -27,10 +27,9 @@ nonisolated enum QuickAddConfiguration {
 
   /// Which built-in preset set a counter should default to, based on its name. Only
   /// "Calories" gets calorie-appropriate presets; every other counter gets the generic set.
-  /// Centralized so a Calories counter looks the same regardless of which surface (settings,
-  /// main page, widget) is the one filling in its defaults — this used to be a settings-only
-  /// check, so the main page and widget silently used generic presets for it instead.
-  static func defaultPresets(forCounterNamed name: String) -> [Int] {
+  /// Prefer `CounterTemplate` when creating new counters — this remains for counters that
+  /// were renamed to "Calories" or migrated from the legacy calorie model.
+  static func defaultPresets(forCounterNamed name: String) -> [Double] {
     name.lowercased() == "calories" ? defaultCaloriePresets : defaultCounterPresets
   }
 
@@ -38,7 +37,7 @@ nonisolated enum QuickAddConfiguration {
   /// user-stored `values`, appends `new` if there's room and `old` wasn't stored (i.e. the
   /// edited slot was only visible via `filledPresets`' fallback), then re-normalizes.
   /// Non-positive input is silently ignored, matching every other preset/amount field.
-  static func replacingPreset(_ old: Int, with new: Int, in values: [Int]) -> [Int] {
+  static func replacingPreset(_ old: Double, with new: Double, in values: [Double]) -> [Double] {
     guard new > 0 else { return values }
 
     var updated = values
